Only install signal handlers in grab mode
It's not necessary to install signal handlers for SIGINT and SIGTERM
when not in grab mode, as there is nothing for evtest to do when
exiting, such as releasing the grab.

This causes problems with umockdev's event capture which relies on being
able to catch the SIGINT signal itself to know to flush its capture
buffers before exiting.

See https://github.com/martinpitt/umockdev/issues/96