Remove PATH_MAX
This makes evtest more portable.  We could try to use ifdefs to include
sys/syslimits.h on FreeBSD to get PATH_MAX but it feels to be an
overkill for evtest.  Let's keep it simple as just replace PATH_MAX with
4096.

NB: it's fine to leave linux/input.h there, because this header is
actually available on FreeBSD via the devel/evdev-proto port.